Add wathc_percentage to the watched movies and series.
Using this new column on the channels table, the developer can register wether the movie has been watched or not.

Currently the string format of the exiting point is being read, and the actual percentage will be calculated next.

The 'completion' statues will be defined, how they will be displayed in another commit.